+++ Runtime/HairSim.cs
@@ -112,8 +112,7 @@ namespace Unity.DemoTeam.Hair
 
 		static class VolumeKernels
 		{
-			public static int KBoundariesAdvance;
-			public static int KBoundariesSubstep;
+			// bounds
 			public static int KBoundsClear;
 			public static int KBoundsGather;
 			public static int KBoundsResolve;
@@ -121,6 +120,8 @@ namespace Unity.DemoTeam.Hair
 			public static int KBoundsHistory;
 			public static int KBoundsCoverage;
 			public static int KLODSelection;
+
+			// grid
 			public static int KVolumeClear;
 			public static int KVolumeSplat;
 			public static int KVolumeSplatDensity;
@@ -137,6 +138,12 @@ namespace Unity.DemoTeam.Hair
 			public static int KVolumeScattering;
 			public static int KVolumeWindPrep;
 			public static int KVolumeWind;
+
+			// scene
+			public static int KBoundariesAdvance;
+			public static int KBoundariesSubstep;
+			public static int KEmittersAdvance;
+			public static int KEmittersSubstep;
 		}
 
 		//TODO move to conf
@@ -292,7 +299,6 @@ namespace Unity.DemoTeam.Hair
 
 				changed |= CreateBuffer(ref volumeBuffers._LODFrustum, "LODFrustum", Conf.MAX_FRUSTUMS, sizeof(LODFrustum));
 
-				changed |= CreateBuffer(ref volumeBuffers._BoundaryRemap, "BoundaryRemap", Conf.MAX_BOUNDARIES, sizeof(int));
 				changed |= CreateBuffer(ref volumeBuffers._BoundaryMatrixNext, "BoundaryMatrixNext", Conf.MAX_BOUNDARIES, sizeof(Matrix4x4));
 				changed |= CreateBuffer(ref volumeBuffers._BoundaryMatrixPrevA, "BoundaryMatrixPrevA", Conf.MAX_BOUNDARIES, sizeof(Matrix4x4));
 				changed |= CreateBuffer(ref volumeBuffers._BoundaryMatrixPrevQ, "BoundaryMatrixPrevQ", Conf.MAX_BOUNDARIES, sizeof(Quaternion));
@@ -300,19 +306,19 @@ namespace Unity.DemoTeam.Hair
 				changed |= CreateBuffer(ref volumeBuffers._BoundaryMatrixInv, "BoundaryMatrixInv", Conf.MAX_BOUNDARIES, sizeof(Matrix4x4));
 				changed |= CreateBuffer(ref volumeBuffers._BoundaryMatrixInvStep, "BoundaryMatrixInvStep", Conf.MAX_BOUNDARIES, sizeof(Matrix4x4));
 				changed |= CreateBuffer(ref volumeBuffers._BoundaryShapeNext, "BoundaryShapeNext", Conf.MAX_BOUNDARIES, sizeof(HairBoundary.RuntimeShape.Data));
+				changed |= CreateBuffer(ref volumeBuffers._BoundaryShapePrevLUT, "BoundaryShapePrevLUT", Conf.MAX_BOUNDARIES, sizeof(int));
 				changed |= CreateBuffer(ref volumeBuffers._BoundaryShapePrev, "BoundaryShapePrev", Conf.MAX_BOUNDARIES, sizeof(HairBoundary.RuntimeShape.Data));
 				changed |= CreateBuffer(ref volumeBuffers._BoundaryShape, "BoundaryShape", Conf.MAX_BOUNDARIES, sizeof(HairBoundary.RuntimeShape.Data));
 
-				//TODO
-				//changed |= CreateBuffer(ref volumeBuffers._WindEmitterRemap, "WindEmitterRemap", Conf.MAX_EMITTERS, sizeof(int));
-				//changed |= CreateBuffer(ref volumeBuffers._WindEmitterDataNext, "WindEmitterDataNext", Conf.MAX_EMITTERS, sizeof(HairWind.RuntimeEmitter));
-				//changed |= CreateBuffer(ref volumeBuffers._WindEmitterDataPrev, "WindEmitterDataPrev", Conf.MAX_EMITTERS, sizeof(HairWind.RuntimeEmitter));
-				//changed |= CreateBuffer(ref volumeBuffers._WindEmitterData, "WindEmitterData", Conf.MAX_EMITTERS, sizeof(HairWind.RuntimeEmitter));
+				changed |= CreateBuffer(ref volumeBuffers._WindEmitterNext, "WindEmitterNext", Conf.MAX_EMITTERS, sizeof(HairWind.RuntimeEmitter));
+				changed |= CreateBuffer(ref volumeBuffers._WindEmitterPrevLUT, "WindEmitterPrevLUT", Conf.MAX_EMITTERS, sizeof(int));
+				changed |= CreateBuffer(ref volumeBuffers._WindEmitterPrev, "WindEmitterPrev", Conf.MAX_EMITTERS, sizeof(HairWind.RuntimeEmitter));
 				changed |= CreateBuffer(ref volumeBuffers._WindEmitter, "WindEmitter", Conf.MAX_EMITTERS, sizeof(HairWind.RuntimeEmitter));
 
 				changed |= CreateBuffer(ref volumeBuffers._BoundsMinMaxU, "BoundsMinMaxU", boundsCount * 2, particleStrideVector3);
-				changed |= CreateBuffer(ref volumeBuffers._Bounds, "Bounds", boundsCount, sizeof(LODBounds));
 				changed |= CreateBuffer(ref volumeBuffers._BoundsPrev, "BoundsPrev", boundsCount, sizeof(LODBounds));
+				changed |= CreateBuffer(ref volumeBuffers._Bounds, "Bounds", boundsCount, sizeof(LODBounds));
+
 				changed |= CreateBuffer(ref volumeBuffers._BoundsGeometry, "BoundsGeometry", boundsCount, sizeof(LODGeometry));
 				changed |= CreateBuffer(ref volumeBuffers._BoundsCoverage, "BoundsCoverage", boundsCount, particleStrideVector2);
 
@@ -421,7 +427,6 @@ namespace Unity.DemoTeam.Hair
 
 			ReleaseBuffer(ref volumeBuffers._LODFrustum);
 
-			ReleaseBuffer(ref volumeBuffers._BoundaryRemap);
 			ReleaseBuffer(ref volumeBuffers._BoundaryMatrixNext);
 			ReleaseBuffer(ref volumeBuffers._BoundaryMatrixPrevA);
 			ReleaseBuffer(ref volumeBuffers._BoundaryMatrixPrevQ);
@@ -429,14 +434,18 @@ namespace Unity.DemoTeam.Hair
 			ReleaseBuffer(ref volumeBuffers._BoundaryMatrixInv);
 			ReleaseBuffer(ref volumeBuffers._BoundaryMatrixInvStep);
 			ReleaseBuffer(ref volumeBuffers._BoundaryShapeNext);
+			ReleaseBuffer(ref volumeBuffers._BoundaryShapePrevLUT);
 			ReleaseBuffer(ref volumeBuffers._BoundaryShapePrev);
 			ReleaseBuffer(ref volumeBuffers._BoundaryShape);
 
+			ReleaseBuffer(ref volumeBuffers._WindEmitterNext);
+			ReleaseBuffer(ref volumeBuffers._WindEmitterPrevLUT);
+			ReleaseBuffer(ref volumeBuffers._WindEmitterPrev);
 			ReleaseBuffer(ref volumeBuffers._WindEmitter);
 
 			ReleaseBuffer(ref volumeBuffers._BoundsMinMaxU);
-			ReleaseBuffer(ref volumeBuffers._Bounds);
 			ReleaseBuffer(ref volumeBuffers._BoundsPrev);
+			ReleaseBuffer(ref volumeBuffers._Bounds);
 			ReleaseBuffer(ref volumeBuffers._BoundsGeometry);
 			ReleaseBuffer(ref volumeBuffers._BoundsCoverage);
 
@@ -473,10 +482,9 @@ namespace Unity.DemoTeam.Hair
 			ReleaseReadbackBuffer(ref volumeData.buffersReadback._BoundsCoverage);
 			ReleaseReadbackBuffer(ref volumeData.buffersReadback._VolumeLODStage);
 
-			if (volumeData.boundaryPrevHandle.IsCreated)
-				volumeData.boundaryPrevHandle.Dispose();
-			if (volumeData.boundaryPrevMatrix.IsCreated)
-				volumeData.boundaryPrevMatrix.Dispose();
+			ReleaseCPUBuffer(ref volumeData.boundaryPrevHandle);
+			ReleaseCPUBuffer(ref volumeData.boundaryPrevMatrix);
+			ReleaseCPUBuffer(ref volumeData.emitterPrevHandle);
 
 			volumeData = new VolumeData();
 		}
@@ -552,7 +560,6 @@ namespace Unity.DemoTeam.Hair
 
 			target.BindComputeBuffer(volumeBufferIDs._LODFrustum, volumeBuffers._LODFrustum);
 
-			target.BindComputeBuffer(volumeBufferIDs._BoundaryRemap, volumeBuffers._BoundaryRemap);
 			target.BindComputeBuffer(volumeBufferIDs._BoundaryMatrixNext, volumeBuffers._BoundaryMatrixNext);
 			target.BindComputeBuffer(volumeBufferIDs._BoundaryMatrixPrevA, volumeBuffers._BoundaryMatrixPrevA);
 			target.BindComputeBuffer(volumeBufferIDs._BoundaryMatrixPrevQ, volumeBuffers._BoundaryMatrixPrevQ);
@@ -560,14 +567,18 @@ namespace Unity.DemoTeam.Hair
 			target.BindComputeBuffer(volumeBufferIDs._BoundaryMatrixInv, volumeBuffers._BoundaryMatrixInv);
 			target.BindComputeBuffer(volumeBufferIDs._BoundaryMatrixInvStep, volumeBuffers._BoundaryMatrixInvStep);
 			target.BindComputeBuffer(volumeBufferIDs._BoundaryShapeNext, volumeBuffers._BoundaryShapeNext);
+			target.BindComputeBuffer(volumeBufferIDs._BoundaryShapePrevLUT, volumeBuffers._BoundaryShapePrevLUT);
 			target.BindComputeBuffer(volumeBufferIDs._BoundaryShapePrev, volumeBuffers._BoundaryShapePrev);
 			target.BindComputeBuffer(volumeBufferIDs._BoundaryShape, volumeBuffers._BoundaryShape);
 
+			target.BindComputeBuffer(volumeBufferIDs._WindEmitterNext, volumeBuffers._WindEmitterNext);
+			target.BindComputeBuffer(volumeBufferIDs._WindEmitterPrevLUT, volumeBuffers._WindEmitterPrevLUT);
+			target.BindComputeBuffer(volumeBufferIDs._WindEmitterPrev, volumeBuffers._WindEmitterPrev);
 			target.BindComputeBuffer(volumeBufferIDs._WindEmitter, volumeBuffers._WindEmitter);
 
 			target.BindComputeBuffer(volumeBufferIDs._BoundsMinMaxU, volumeBuffers._BoundsMinMaxU);
-			target.BindComputeBuffer(volumeBufferIDs._Bounds, volumeBuffers._Bounds);
 			target.BindComputeBuffer(volumeBufferIDs._BoundsPrev, volumeBuffers._BoundsPrev);
+			target.BindComputeBuffer(volumeBufferIDs._Bounds, volumeBuffers._Bounds);
 			target.BindComputeBuffer(volumeBufferIDs._BoundsGeometry, volumeBuffers._BoundsGeometry);
 			target.BindComputeBuffer(volumeBufferIDs._BoundsCoverage, volumeBuffers._BoundsCoverage);
 
@@ -1379,7 +1390,7 @@ namespace Unity.DemoTeam.Hair
 		{
 			ref var volumeConstantsEnvironment = ref volumeData.constantsEnvironment;
 
-			// capture frustums
+			// update frustums
 			using (var frustums = AcquireLODFrustums(CameraType.Game | CameraType.SceneView, Allocator.Temp))
 			{
 				//int i = 0;
@@ -1398,7 +1409,7 @@ namespace Unity.DemoTeam.Hair
 				//	Debug.Log("frustum.plane5 = " + frustum.plane5);
 				//}
 
-				// derive constants
+				// write constants
 				volumeConstantsEnvironment._LODFrustumCount = (uint)frustums.Length;
 
 				// update buffers
@@ -1414,10 +1425,10 @@ namespace Unity.DemoTeam.Hair
 			ref var volumeConstantsScene = ref volumeData.constantsEnvironment;
 			ref var volumeTextures = ref volumeData.textures;
 
-			// capture gravity
+			// update gravity
 			volumeConstantsScene._WorldGravity = settingsEnvironment.gravityRotation * (Physics.gravity * settingsEnvironment.gravityScale);
 
-			// capture boundaries
+			// update boundaries
 			using (var bufRemap = new NativeArray<int>(Conf.MAX_BOUNDARIES, Allocator.Temp, NativeArrayOptions.ClearMemory))
 			using (var bufHandle = new NativeArray<int>(Conf.MAX_BOUNDARIES, Allocator.Temp, NativeArrayOptions.ClearMemory))
 			using (var bufMatrix = new NativeArray<Matrix4x4>(Conf.MAX_BOUNDARIES, Allocator.Temp, NativeArrayOptions.ClearMemory))
@@ -1425,15 +1436,8 @@ namespace Unity.DemoTeam.Hair
 			using (var bufMatrixQ = new NativeArray<Quaternion>(Conf.MAX_BOUNDARIES, Allocator.Temp, NativeArrayOptions.ClearMemory))
 			using (var bufShape = new NativeArray<HairBoundary.RuntimeShape.Data>(Conf.MAX_BOUNDARIES, Allocator.Temp, NativeArrayOptions.ClearMemory))
 			{
-				if (volumeData.boundaryPrevHandle.IsCreated && volumeData.boundaryPrevHandle.Length != Conf.MAX_BOUNDARIES)
-					volumeData.boundaryPrevHandle.Dispose();
-				if (volumeData.boundaryPrevHandle.IsCreated == false)
-					volumeData.boundaryPrevHandle = new NativeArray<int>(Conf.MAX_BOUNDARIES, Allocator.Persistent, NativeArrayOptions.ClearMemory);
-
-				if (volumeData.boundaryPrevMatrix.IsCreated && volumeData.boundaryPrevMatrix.Length != Conf.MAX_BOUNDARIES)
-					volumeData.boundaryPrevMatrix.Dispose();
-				if (volumeData.boundaryPrevMatrix.IsCreated == false)
-					volumeData.boundaryPrevMatrix = new NativeArray<Matrix4x4>(Conf.MAX_BOUNDARIES, Allocator.Persistent, NativeArrayOptions.ClearMemory);
+				CreateCPUBuffer(ref volumeData.boundaryPrevHandle, Conf.MAX_BOUNDARIES, Allocator.Persistent, NativeArrayOptions.ClearMemory);
+				CreateCPUBuffer(ref volumeData.boundaryPrevMatrix, Conf.MAX_BOUNDARIES, Allocator.Persistent, NativeArrayOptions.ClearMemory);
 
 				unsafe
 				{
@@ -1446,7 +1450,7 @@ namespace Unity.DemoTeam.Hair
 
 					// gather boundaries
 					//TODO expose or always enable the volumeSort option which sorts active boundaries by distance
-					var boundaryList = HairBoundaryUtility.Gather(settingsEnvironment.boundaryResident, settingsEnvironment.boundaryCapture, GetVolumeBounds(volumeData), settingsEnvironment.boundaryCaptureLayer, captureSort: false, (settingsEnvironment.boundaryCaptureMode == SettingsEnvironment.BoundaryCaptureMode.IncludeColliders));
+					var boundaryList = SpatialComponentFilter<HairBoundary, HairBoundary.RuntimeData, HairBoundaryProxy>.Gather(settingsEnvironment.boundaryResident, settingsEnvironment.boundaryCapture, GetVolumeBounds(volumeData), settingsEnvironment.boundaryCaptureLayer, volumeSort: false, (settingsEnvironment.boundaryCaptureMode == SettingsEnvironment.BoundaryCaptureMode.IncludeColliders));
 
 					var boundaryCountDiscrete = 0;
 					var boundaryCountCapsule = 0;
@@ -1496,7 +1500,7 @@ namespace Unity.DemoTeam.Hair
 					var firstIndexCube = firstIndexTorus + boundaryCountTorus;
 					var firstIndexEnd = firstIndexCube + boundaryCountCube;
 
-					// update constants
+					// write constants
 					volumeConstantsScene._BoundaryDelimDiscrete = (uint)firstIndexCapsule;
 					volumeConstantsScene._BoundaryDelimCapsule = (uint)firstIndexSphere;
 					volumeConstantsScene._BoundaryDelimSphere = (uint)firstIndexTorus;
@@ -1506,11 +1510,8 @@ namespace Unity.DemoTeam.Hair
 					volumeConstantsScene._BoundaryWorldEpsilon = (boundarySDFIndex != -1) ? boundarySDFCellSize * 0.2f : 1e-4f;
 					volumeConstantsScene._BoundaryWorldMargin = settingsEnvironment.defaultSolidMargin * 0.01f;
 
-					// assign discrete texture
-					if (boundarySDFIndex != -1)
-					{
-						volumeTextures._BoundarySDF = (boundarySDFIndex != -1) ? boundaryList[boundarySDFIndex].sdf.sdfTexture : null;
-					}
+					// write textures
+					volumeTextures._BoundarySDF = (boundarySDFIndex != -1) ? boundaryList[boundarySDFIndex].sdf.sdfTexture : null;
 
 					// write attributes
 					{
@@ -1537,7 +1538,7 @@ namespace Unity.DemoTeam.Hair
 							}
 							else
 							{
-								if (writeIndexDiscrete == 0)// only include one first discrete
+								if (writeIndexDiscrete == 0)// only include first discrete
 								{
 									writeIndex = writeIndexDiscrete++;
 								}
@@ -1603,11 +1604,11 @@ namespace Unity.DemoTeam.Hair
 					}
 
 					// update buffers
-					PushComputeBufferData(cmd, volumeData.buffers._BoundaryRemap, bufRemap);
 					PushComputeBufferData(cmd, volumeData.buffers._BoundaryMatrixNext, bufMatrix);
 					PushComputeBufferData(cmd, volumeData.buffers._BoundaryMatrixPrevA, bufMatrixA);
 					PushComputeBufferData(cmd, volumeData.buffers._BoundaryMatrixPrevQ, bufMatrixQ);
 					PushComputeBufferData(cmd, volumeData.buffers._BoundaryShapeNext, bufShape);
+					PushComputeBufferData(cmd, volumeData.buffers._BoundaryShapePrevLUT, bufRemap);
 
 					// update previous frame info
 					if (stepCount > 0)
@@ -1648,50 +1649,67 @@ namespace Unity.DemoTeam.Hair
 				}
 			}
 
-			// capture emitters
+			// update emitters
+			using (var bufRemap = new NativeArray<int>(Conf.MAX_EMITTERS, Allocator.Temp, NativeArrayOptions.ClearMemory))
+			using (var bufHandle = new NativeArray<int>(Conf.MAX_EMITTERS, Allocator.Temp, NativeArrayOptions.ClearMemory))
 			using (var bufEmitter = new NativeArray<HairWind.RuntimeEmitter>(Conf.MAX_EMITTERS, Allocator.Temp, NativeArrayOptions.ClearMemory))
 			{
+				CreateCPUBuffer(ref volumeData.emitterPrevHandle, Conf.MAX_EMITTERS, Allocator.Persistent, NativeArrayOptions.ClearMemory);
+
 				unsafe
 				{
+					var ptrRemap = (int*)bufRemap.GetUnsafePtr();
+					var ptrHandle = (int*)bufHandle.GetUnsafePtr();
 					var ptrEmitter = (HairWind.RuntimeEmitter*)bufEmitter.GetUnsafePtr();
 
-					// derive emitters
-					//TODO move to HairWindUtility
-					var emitterList = new UnsafeList<HairWind.RuntimeData>(Conf.MAX_EMITTERS, Allocator.Temp, NativeArrayOptions.UninitializedMemory);
-					var emitterData = new HairWind.RuntimeData();
+					// gather emitters
+					var emitterList = SpatialComponentFilter<HairWind, HairWind.RuntimeData, HairWindProxy>.Gather(settingsEnvironment.emitterResident, settingsEnvironment.emitterCapture, GetVolumeBounds(volumeData), settingsEnvironment.emitterCaptureLayer, volumeSort: false, (settingsEnvironment.emitterCaptureMode == SettingsEnvironment.EmitterCaptureMode.IncludeWindZones));
+					var emitterCount = Mathf.Min(Conf.MAX_EMITTERS, emitterList.Count);
 
-					foreach (var emitter in HairWind.s_emitters)
+					// write constants
+					volumeConstantsScene._WindEmitterCount = (uint)emitterCount;
+					volumeConstantsScene._WindEmitterClock = (float)elapsedTime;
+
+					// write attributes
 					{
-						if (emitter == null || emitter.isActiveAndEnabled == false)
-							continue;
-
-						if (HairWind.TryGetData(emitter, ref emitterData))
+						for (int i = 0; i != emitterCount; i++)
 						{
-							emitterList.Add(emitterData);
+							ptrEmitter[i] = emitterList[i].emitter;
 						}
 					}
 
-					// derive constants
-					volumeConstantsScene._WindEmitterCount = (uint)Mathf.Min(Conf.MAX_EMITTERS, emitterList.Length);
-					volumeConstantsScene._WindEmitterClock = (float)elapsedTime;
-
-					// update emitter data
-					for (int i = 0; i != volumeConstantsScene._WindEmitterCount; i++)
+					// write remapping (current -> previous)
 					{
-						ptrEmitter[i] = emitterList[i].emitter;
+						for (int i = 0; i != emitterCount; i++)
+						{
+							ptrRemap[i] = volumeData.emitterPrevHandle.IndexOf(ptrHandle[i]);
+						}
+
+						for (int i = emitterCount; i != Conf.MAX_EMITTERS; i++)
+						{
+							ptrRemap[i] = -1;
+						}
 					}
 
-					emitterList.Dispose();
-
 					// update buffers
-					PushComputeBufferData(cmd, volumeData.buffers._WindEmitter, bufEmitter);
+					PushComputeBufferData(cmd, volumeData.buffers._WindEmitterNext, bufEmitter);
+					PushComputeBufferData(cmd, volumeData.buffers._WindEmitterPrevLUT, bufRemap);
+
+					// update previous frame info
+					if (stepCount > 0)
+					{
+						volumeData.emitterPrevHandle.CopyFrom(bufHandle);
+					}
+
+					volumeData.emitterCount = emitterCount;
+					volumeData.emitterCountDiscard = emitterList.Count - emitterCount;
 				}
 			}
 
 			// update cbuffer
 			PushConstantBufferData(cmd, volumeData.buffers.VolumeCBufferEnvironment, volumeConstantsScene);
 
-			// advance boundaries
+			// conditionally advance boundaries
 			if (stepCount > 0)
 			{
 				int numX = (Conf.MAX_BOUNDARIES + THREAD_GROUP_SIZE - 1) / THREAD_GROUP_SIZE;
@@ -1700,6 +1718,17 @@ namespace Unity.DemoTeam.Hair
 
 				BindVolumeData(cmd, s_volumeCS, VolumeKernels.KBoundariesAdvance, volumeData);
 				cmd.DispatchCompute(s_volumeCS, VolumeKernels.KBoundariesAdvance, numX, numY, numZ);
+			}
+
+			// conditionally advance emitters
+			if (stepCount > 0)
+			{
+				int numX = (Conf.MAX_EMITTERS + THREAD_GROUP_SIZE - 1) / THREAD_GROUP_SIZE;
+				int numY = 1;
+				int numZ = 1;
+
+				BindVolumeData(cmd, s_volumeCS, VolumeKernels.KEmittersAdvance, volumeData);
+				cmd.DispatchCompute(s_volumeCS, VolumeKernels.KEmittersAdvance, numX, numY, numZ);
 			}
 		}
 
@@ -1771,12 +1800,12 @@ namespace Unity.DemoTeam.Hair
 			{
 				if (stepFracHi > 0.0f)
 				{
-					// substep boundaries
 					//TODO skip if at fraction one, similar to roots
-					{
-						cmd.SetComputeFloatParam(s_volumeCS, "_SubstepFractionLo", stepFracLo);
-						cmd.SetComputeFloatParam(s_volumeCS, "_SubstepFractionHi", stepFracHi);
+					cmd.SetComputeFloatParam(s_volumeCS, "_SubstepFractionLo", stepFracLo);
+					cmd.SetComputeFloatParam(s_volumeCS, "_SubstepFractionHi", stepFracHi);
 
+					// substep boundaries
+					{
 						int numX = (Conf.MAX_BOUNDARIES + THREAD_GROUP_SIZE - 1) / THREAD_GROUP_SIZE;
 						int numY = 1;
 						int numZ = 1;
@@ -1786,7 +1815,14 @@ namespace Unity.DemoTeam.Hair
 					}
 
 					// substep emitters
-					//TODO
+					{
+						int numX = (Conf.MAX_EMITTERS + THREAD_GROUP_SIZE - 1) / THREAD_GROUP_SIZE;
+						int numY = 1;
+						int numZ = 1;
+
+						BindVolumeData(cmd, s_volumeCS, VolumeKernels.KEmittersSubstep, volumeData);
+						cmd.DispatchCompute(s_volumeCS, VolumeKernels.KEmittersSubstep, numX, numY, numZ);
+					}
 				}
 
 				HairSim.PushVolumeClear(cmd, ref volumeData, settingsVolume);